Accurately parsed tables using colspans
Okay: for the purpose of the journal:
We figured out the only consistency was colspans, but then an issue popped up where some tags didn't have the attribute. We figured out that the number of td's without the attribute just happened to be equal to 3, which was the value for all other colspans. So when the tag wasn't there, we said span = 1.

Oh, and we removed display nones. Those are there for no reason, and sometimes have colspan values greater than one.

The code was able to accurately parse a table.